requirements.txt
streamlit
tabula-py
pandas
xlsxwriter